adds new incomes controller, and views

+++ db/migrate/20150418000855_create_debts.rb
@@ -2,9 +2,9 @@ class CreateDebts < ActiveRecord::Migration
   def change
     create_table :debts do |t|
     	t.string :type
-    	t.decimal :amount
-    	t.decimal :monthly_payment
-    	t.decimal :rate
+    	t.float :amount
+    	t.float :monthly_payment
+    	t.float :rate
     	t.string :act_primary
     	t.string :act_secondary
